requirements.txt
requests~=2.32.1
python-dotenv~=1.0.1